Change empty state location
