+++ src/components/Errors/ErrorsList/styles.ts
@@ -1,8 +1,5 @@
 import styled from "styled-components";
-import {
-  bodySemiboldTypography,
-  footnoteRegularTypography
-} from "../../common/App/typographies";
+import { bodySemiboldTypography } from "../../common/App/typographies";
 
 export const Container = styled.div`
   display: flex;
@@ -10,16 +7,6 @@ export const Container = styled.div`
   gap: 8px;
   padding: 8px;
   overflow: auto;
-`;
-
-export const EmptyStateContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-  justify-content: center;
-  flex-grow: 1;
-  align-self: center;
 `;
 
 export const EmptyStateIconContainer = styled.div`
@@ -31,18 +18,6 @@ export const EmptyStateIconContainer = styled.div`
   border-radius: 50%;
   color: ${({ theme }) => theme.colors.v3.icon.secondary};
   background: ${({ theme }) => theme.colors.v3.surface.sidePanelHeader};
-`;
-
-export const EmptyStateTextContainer = styled.div`
-  ${footnoteRegularTypography}
-
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  gap: 4px;
-  padding-top: 4px;
-  padding-bottom: 4px;
-  color: ${({ theme }) => theme.colors.v3.text.tertiary};
 `;
 
 export const EmptyStateTitle = styled.div`
